adding functions to register a user in the database and read user information
